move delete user into its own class
